dev(dataset): change the method 'open'
Now the method uses shareable global lock, but  when need to do initialization of a dataset there is a gateway with scoped lock of key_table.
This gateway helps to initialize a dataset by several processes and only one process really  does initialization of the dataset. This optimization helps to reduce time of opening a dataset for several processes.